@@ -228,7 +228,7 @@ func (csm *Manager) UploadHeadImg(kfAccount, fileName string) (err error) {
 	return
 }
 
-//SendTypingStatus 下发客服输入状态给用户
+// SendTypingStatus 下发客服输入状态给用户
 func (csm *Manager) SendTypingStatus(openid string, cmd TypingStatus) (err error) {
 	var accessToken string
 	accessToken, err = csm.GetAccessToken()
